requirements.txt
langchain
python-dotenv
langchain-community
azure-ai-documentintelligence
markdown
beautifulsoup4
